% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/float.R
\name{floatFUN}
\alias{floatFUN}
\title{Two-column creator}
\usage{
floatFUN(leftList, rightList, width)
}
\arguments{
\item{leftList}{the list for all visuals on the left column}

\item{rightList}{the list for all visuals on the right column}

\item{width}{a two-numbered vector for the relative width of left and right (in percent)}
}
\value{
a two-column layout
}
\description{
Create a two-column layout for lists of different visuals
}
\examples{
two_column_vis <- floatFUN(list(table1),list(plot1),c(40,60))
}
