% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modalFUN}
\alias{modalFUN}
\title{Modal creator}
\usage{
modalFUN(visualList, buttonLabel)
}
\arguments{
\item{visualList}{the list of all visuals to put in modal}

\item{buttonLabel}{label to give button that activates modal}
}
\value{
an action button that, when pressed, presents a modal
}
\description{
Create a modal (shown via action button) for a list of visuals
}
\examples{
modal_button <- modalFUN(list(plot1),"Click for modal")
}
