% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabFUN}
\alias{tabFUN}
\title{Tab navigation creator}
\usage{
tabFUN(visualList, tabVec)
}
\arguments{
\item{visualList}{the list of all visuals}

\item{tabVec}{a vector of character names for each visual of the tabs}
}
\value{
a tab navigation output
}
\description{
Create a tab navigator for a list of visuals
}
\examples{
tab_vis <- tabFUN(list(table1,plot1),c("Table","Plot"))
}
